requirements_colab.txt
aiohttp==3.9.0
APScheduler==3.10.4
arxiv==2.0.0
beautifulsoup4==4.12.3
charset_normalizer==3.3.2
class_registry==2.1.2
docx==0.2.4
fastapi==0.109.2
func_timeout==4.3.5
google_api_python_client==2.111.0
gradio==4.18.0
httpx==0.26.0
jsonlines==4.0.0
loguru==0.7.2
Markdown==3.5.1
networkx==3.2.1
openai==1.12.0
opencv_python==4.8.1.78
openpyxl==3.1.2
pandas==1.5.3
pydantic==2.6.1
pylatexenc==2.10
PyPDF2==3.0.1
python-dotenv==1.0.1
python_docx==1.1.0
python_pptx==0.6.23
pytube==15.0.0
pyvis==0.3.2
PyYAML==6.0.1
PyYAML==6.0.1
regex==2023.10.3
Requests==2.31.0
scipy==1.12.0
seaborn==0.13.2
sentence_transformers==2.2.2
shortuuid==1.0.11
tenacity==8.2.3
termcolor==2.4.0
tqdm==4.66.1
transformers==4.36.2
wikipedia==1.4.0
